+++ src/main/scala/ai/privado/exporter/DataflowExporter.scala
@@ -107,7 +107,7 @@ class DataflowExporter(cpg: Cpg, dataflowsMap: Map[String, Path]) {
     sourceModelList.foreach(sourceModel => {
       var sinkId = sourceModel.sinkId
       val sinkAPITag = dataflowsMap(sourceModel.pathId).elements.last.tag
-        .filter(node => node.name.equals(Constants.apiUrl))
+        .filter(node => node.name.equals(Constants.apiUrl + sinkSubCategory))
       if (sinkAPITag.nonEmpty) {
         sinkId += "#_#" + sinkAPITag.value.l.mkString(",")
       }
