+++ src/main/scala/ai/privado/languageEngine/java/tagger/sink/JavaAPITagger.scala
@@ -23,26 +23,17 @@
 package ai.privado.languageEngine.java.tagger.sink
 
 import ai.privado.cache.{AppCache, RuleCache}
-import ai.privado.languageEngine.java.language.{NodeStarters, NodeToProperty, StepsForProperty}
+import ai.privado.languageEngine.java.language.{NodeStarters, StepsForProperty}
 import ai.privado.metric.MetricHandler
-import ai.privado.model.{Constants, Language, NodeType, RuleInfo, SystemConfig}
+import ai.privado.model.{Language, NodeType, RuleInfo}
+import ai.privado.tagger.utility.APITaggerUtility.sinkTagger
 import ai.privado.utility.ImportUtility
-import ai.privado.utility.Utilities.{
-  addRuleTags,
-  getDefaultSemantics,
-  getFileNameForNode,
-  isFileProcessable,
-  storeForTag
-}
 import io.circe.Json
-import io.joern.dataflowengineoss.language._
-import io.joern.dataflowengineoss.queryengine.{EngineConfig, EngineContext}
 import io.shiftleft.codepropertygraph.generated.Cpg
-import io.shiftleft.codepropertygraph.generated.nodes.{Call, CfgNode}
+import io.shiftleft.codepropertygraph.generated.nodes.Call
 import io.shiftleft.passes.ForkJoinParallelCpgPass
 import io.shiftleft.semanticcpg.language._
 import org.slf4j.LoggerFactory
-import overflowdb.BatchedUpdate
 
 import java.util.Calendar
 import scala.collection.mutable
@@ -107,8 +98,7 @@ class JavaAPITagger(cpg: Cpg) extends ForkJoinParallelCpgPass[RuleInfo](cpg) {
       APITaggerVersionJava.V2Tagger
   }
 
-  implicit val engineContext: EngineContext = EngineContext(semantics = getDefaultSemantics, config = EngineConfig(4))
-  val systemConfigHttpLibraries             = RuleCache.getSystemConfigByKey("apiHttpLibraries")
+  val systemConfigHttpLibraries = RuleCache.getSystemConfigByKey("apiHttpLibraries")
   val commonHttpPackages: String = {
     systemConfigHttpLibraries.size match {
       case 0 =>
@@ -140,31 +130,6 @@ class JavaAPITagger(cpg: Cpg) extends ForkJoinParallelCpgPass[RuleInfo](cpg) {
       case _ =>
         logger.debug("Skipping API Tagger because valid match not found")
         println(s"${Calendar.getInstance().getTime} - --API TAGGER SKIPPED...")
-    }
-  }
-
-  private def sinkTagger(
-    apiInternalSinkPattern: List[CfgNode],
-    apis: List[CfgNode],
-    builder: BatchedUpdate.DiffGraphBuilder,
-    ruleInfo: RuleInfo
-  ): Unit = {
-    val filteredLiteralSourceNode = apiInternalSinkPattern.filter(node => isFileProcessable(getFileNameForNode(node)))
-    apis.foreach(apiNode => addRuleTags(builder, apiNode, ruleInfo))
-    if (apis.nonEmpty && filteredLiteralSourceNode.nonEmpty) {
-      val apiFlows = apis.reachableByFlows(filteredLiteralSourceNode).l
-      apiFlows.foreach(flow => {
-        val literalCode = flow.elements.head.originalPropertyValue.getOrElse(flow.elements.head.code)
-        val apiNode     = flow.elements.last
-        // addRuleTags(builder, apiNode, ruleInfo)
-        storeForTag(builder, apiNode)(Constants.apiUrl, literalCode)
-      })
-      val literalPathApiNodes        = apiFlows.map(_.elements.last).toSet
-      val apiNodesWithoutLiteralPath = apis.toSet.diff(literalPathApiNodes)
-      apiNodesWithoutLiteralPath.foreach(apiNode => storeForTag(builder, apiNode)(Constants.apiUrl, "https://API.com"))
-    }
-    if (filteredLiteralSourceNode.isEmpty) {
-      apis.foreach(apiNode => storeForTag(builder, apiNode)(Constants.apiUrl, "https://API.com"))
     }
   }
