+++ src/main/scala/ai/privado/model/Constants.scala
@@ -106,4 +106,8 @@ object Constants {
   val PYTHON_STATS = "python.txt"
 
   val delombok = "delombok"
+  val API      = "API"
+
+  // Rule ids
+  val internalAPIRuleId = "Sinks.API.InternalAPI"
 }
